requirements.txt
fastapi
uvicorn
pydantic-settings
SQLAlchemy
psycopg[binary]
httpx
requests
pandas
numpy
python-dotenv
sentence-transformers
transformers==4.45.2
chromadb
rank-bm25
langchain
langchain-ollama
langchain-openai
langchain-community
openai
FlagEmbedding
faiss-cpu
ragas
